+++ previous-versions/common/RELEASES.md
@@ -39,3 +39,13 @@ Based on:
 - [python v0.36.0] previous-versions/common
 ### Releases
 - [PyPI v0.36.0] https://pypi.org/project/codat-common/0.36.0 - previous-versions/common
+
+## 2023-10-02 00:16:57
+### Changes
+Based on:
+- OpenAPI Doc 3.0.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Platform.yaml
+- Speakeasy CLI 1.92.3 (2.143.2) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v0.36.1] previous-versions/common
+### Releases
+- [PyPI v0.36.1] https://pypi.org/project/codat-common/0.36.1 - previous-versions/common